src/main/java/org/bargains/config/validators/ValidDoubleValidator.java
package org.bargains.config.validators;

import javax.validation.ConstraintValidator;
import javax.validation.ConstraintValidatorContext;
import java.time.format.DateTimeParseException;

public class ValidDoubleValidator implements ConstraintValidator<ValidDouble, String> {

    @Override
    public void initialize(ValidDouble constraintAnnotation) {

    }

    @Override
    public boolean isValid(String value, ConstraintValidatorContext context) {
        if (value == null)
            return true;
        try {
            Double.parseDouble(value);
        } catch (NumberFormatException e) {
            return false;
        }
        return true;
    }
}
